tiendanube_service/requirements.txt
fastapi
uvicorn
pydantic
httpx
tenacity
structlog
python-dotenv
prometheus-client
